handled branch overflow through searching
of the next ancestor with the according orientation
(left, right: horizontal; up, down: vertical)
and sending a ChangeFocus with the path to this ancestor,
but with incremented (right, down) or decremented (left, up) children index,
of the according ancestor.

If this fails (through out of bounds due to the in/decrementation) repeating of this
process (remitting of a ChangeFocus) till it reaches the root node
=> default focus.

The default leave changed also, depending on the move direction so that in this case
(in case the new path (ChangeFocus) is not long enough to reach a leave):

            r
           / \
          n  lf        where "lf" is the focused leave,
         / \           "r" and "n" have the same (horizontal) orientation,
        l1 l2          and we want to move to the left.

we dont focus "l1" like the default would have before (since its the first
in the child array of "n") but to "l2" because its the first in the
direction (assumed that so the last of "n" children's array.