captured children-index underflow
since we know that a negative index is not a valid array index,
we can handle it, before emitting a ChangeFocus with a wrong index (path)
and keep the focus directly and thus don't change unnecessary the focus
to the default (which is the current since its the last focused)
when tried to move out of the screen.

This reduces the complexity of the focus change, so much so that we
should add the index length for each node into the path of a FocusLeave,
to be able to catch the index overflow also.

- added a lastFocused so that:
 +---++---+  move   +---++---+  move   +---++---+
 |   ||   |  right  |   ||   |  left   |   ||   |
 +---+|   | ------> +---+| f | ------> +---+|   |   =>  last equals first
 | f ||   |         |   ||   |         | f ||   |
 +---++---+         +---++---+         +---++---+

and not:
 +---++---+  move   +---++---+  move   +---++---+
 |   ||   |  right  |   ||   |  left   | f ||   |
 +---+|   | ------> +---+| f | ------> +---+|   |   => last does not equals first
 | f ||   |         |   ||   |         |   ||   |      because of default.
 +---++---+         +---++---+         +---++---+
this is recursive true, so each node of the boxer-tree remembers which
child was focus last and if no path information is provided (path to
short) than the lastFocused is given the focus (maybe to give it further
down (recursively)).